etl/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "google-cloud-storage>=3.2.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "sqlmesh[duckdb]>=0.202.1",
]
